fix(android): move screen wakelock to main process to fix video call screen-off (WT-1366)
ActivityWakelockManager lived in :callkeep_core (PhoneConnectionServiceDispatcher) where
ActivityHolder.getActivity() always returns null - FLAG_KEEP_SCREEN_ON was never applied.

- Move ActivityWakelockManager to foreground package (main process, correct JVM)
- Add isScreenOnDesired field so Activity recreation re-applies the flag
- Add updateMetadata() to ConnectionTracker so mid-call video upgrades reach the shadow
- Sync wakelock from ForegroundService at 5 lifecycle points (promote/update/terminate/tearDown)
- Remove dead ActivityWakelockManager from PhoneConnectionServiceDispatcher and PhoneConnectionService
