+++ webtrit_callkeep_android/android/src/main/kotlin/com/webtrit/callkeep/services/services/foreground/ActivityWakelockManager.kt
@@ -1,4 +1,4 @@
-package com.webtrit.callkeep.services.services.connection
+package com.webtrit.callkeep.services.services.foreground
 
 import android.app.Activity
 import android.view.WindowManager
@@ -10,12 +10,19 @@ class ActivityWakelockManager(
 ) {
     private val operationQueue = mutableListOf<(Activity) -> Unit>()
 
+    @Volatile private var isScreenOnDesired = false
+
     // Reference to the listener for unsubscribing later
     private val activityChangeListener: (Activity?) -> Unit = { activity ->
         val activityName = activity?.componentName?.shortClassName ?: "null"
         logger.d("Activity lifecycle change detected. Current Activity: $activityName")
 
-        activity?.let { executePendingOperations(it) }
+        activity?.let {
+            if (isScreenOnDesired) {
+                it.window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+            }
+            executePendingOperations(it)
+        }
     }
 
     init {
@@ -27,6 +34,7 @@ class ActivityWakelockManager(
      * Keeps the screen on by applying the FLAG_KEEP_SCREEN_ON to the current activity.
      */
     fun acquireScreenWakeLock() {
+        isScreenOnDesired = true
         executeOrQueue("Acquire FLAG_KEEP_SCREEN_ON") { activity ->
             logger.v("Applying window flag FLAG_KEEP_SCREEN_ON to ${activity.componentName.shortClassName}")
             activity.window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
@@ -37,6 +45,7 @@ class ActivityWakelockManager(
      * Releases the wake lock by clearing the FLAG_KEEP_SCREEN_ON from the current activity.
      */
     fun releaseScreenWakeLock() {
+        isScreenOnDesired = false
         executeOrQueue("Release FLAG_KEEP_SCREEN_ON") { activity ->
             logger.v("Clearing window flag FLAG_KEEP_SCREEN_ON from ${activity.componentName.shortClassName}")
             activity.window.clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
@@ -99,6 +108,8 @@ class ActivityWakelockManager(
      */
     fun dispose() {
         logger.d("Disposing ActivityWakelockManager. Cleanup started.")
+
+        isScreenOnDesired = false
 
         activityProvider.getActivity()?.let { activity ->
             runCatching {
